+++ src/main/java/sptech/elderly/repository/PropostaRepository.java
@@ -3,5 +3,5 @@ package sptech.elderly.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import sptech.elderly.entity.Proposta;
 
-public interface PropostaRepository extends JpaRepository<Proposta, Integer> {
+public interface PropostaRepository extends JpaRepository<Proposta, Long> {
 }
