+++ src/main/java/sptech/elderly/repository/TipoUsuarioRepository.java
@@ -3,5 +3,5 @@ package sptech.elderly.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import sptech.elderly.entity.TipoUsuario;
 
-public interface TipoUsuarioRepository extends JpaRepository<TipoUsuario, Integer> {
+public interface TipoUsuarioRepository extends JpaRepository<TipoUsuario, Long> {
 }
