+++ src/main/java/sptech/elderly/web/dto/mensagem/MensagemComPropostaInput.java
@@ -5,8 +5,8 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import sptech.elderly.web.dto.proposta.PropostaInput;
 
-public record MensagemComPropostaInput(@NotNull Integer remetenteId,
-                                       @NotNull Integer destinatarioId,
+public record MensagemComPropostaInput(@NotNull Long remetenteId,
+                                       @NotNull Long destinatarioId,
                                        @NotBlank String conteudo,
                                        @NotNull @Valid PropostaInput proposta) {
 }
